Add Pagination to Dashboard Sections and Minor Fixes

+++ products/views.py
@@ -211,14 +211,21 @@ class ProductDetailsView(View):
 
 def ProductCommentDeleteView(request, pk):
     """ Delete a product from the store """
-
     comment = get_object_or_404(ProductComment, id=pk)
-    comment.delete()
+    if request.user.is_authenticated:
+        comment.delete()
 
-    messages.success(
-        request,
-        'Comment deleted successfully!'
-        )
+        messages.success(
+            request,
+            'Comment deleted successfully!'
+            )
+        return redirect(reverse(
+            'product_details',
+            kwargs={'slug': comment.product.slug}))
+    messages.error(
+            request,
+            'Permission Denied! You cannot delete this comment!'
+            )
     return redirect(reverse(
-        'product_details',
-        kwargs={'slug': comment.product.slug}))
+            'product_details',
+            kwargs={'slug': comment.product.slug}))
